requests_ecp.ecp: further refactoring
to standalone `ecp.authenticate()` function
